Move auth and geocoding_worker out of repo root
The last two strays at root (auth.py and geocoding_worker.py) were both
imported as if they were top-level packages — but they belong with
their respective agent features.

- git mv auth.py                agents/auth/credentials.py
- git mv geocoding_worker.py    agents/itinerary/geocoding_worker.py
- agents/auth/routes.py:           import auth → from agents.auth import credentials as auth
- agents/trips/routes.py:          import geocoding_worker → from agents.itinerary import geocoding_worker
- agents/create/handler.py: same
- agents/create/upload_handlers.py (2 lazy imports): same
- agents/pages/routes.py:          from geocoding_worker import … →
  from agents.itinerary.geocoding_worker import …
- scripts/check_file_size.py: drop auth.py / geocoding_worker.py from
  INCLUDE_ROOT_FILES; only app.py remains at root.

Top-level Python is now just app.py. Tests still 317 passing.
