+++ agents/auth/credentials.py
